packages/djsailthru/djsailthru-0.0.7.tar.gz/djsailthru-0.0.7/djsailthru/mail/message.py
from django.core.mail import EmailMessage


class SailthruEmailMessage(EmailMessage):
    def __init__(self, email_address, template_name, connection=None, **kwargs):
        self.template_name = template_name
        self.email_address = email_address
        self.vars = kwargs

        self.data = {
            "template": self.template_name,
            "email": self.email_address,
            "vars": self.vars
        }
        self.connection = connection

    def message(self):
        """
        This allows SailthruEmailMessage to function within Django's
        localmem EmailBackend when running tests
        """
        return self

    def recipients(self):
        return self.email_address

    def as_bytes(self):
        return self.data
